end of 37

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -3,6 +3,7 @@ const jwt = require("jsonwebtoken");
 const { randomBytes } = require("crypto");
 const { promisify } = require("util");
 const { makeANiceEmail, transport } = require("../mail");
+const { hasPermission } = require("../utils");
 
 const Mutation = {
   async createItem(parent, args, ctx, info) {
@@ -49,9 +50,18 @@ const Mutation = {
     const where = { id: args.id };
 
     // 1. find the item
-    const item = ctx.db.query.item({ where }, `{ id title }`);
-    // 2. Check if they have permission
-    // TODO
+    const item = await ctx.db.query.item({ where }, `{ id title user { id } }`);
+    // 2. Check if they have permission or if they own it
+
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(
+      permission => ['ADMIN', 'ITEMDELETE'].includes(permission)
+    );
+
+    if (!ownsItem && !hasPermissions) {
+      throw new Error('You aren\'t allowed!');
+    }
+
     // 3. delete it
     return ctx.db.mutation.deleteItem({ where }, info);
   },
@@ -172,6 +182,34 @@ const Mutation = {
     // 8. return the new user
     return updatedUser;
   },
+  async updatePermissions(parent, args, ctx, info) {
+    // 1. Check if they are logged in
+    if (!ctx.request.userId) {
+      throw new Error('You must be logged in!');
+    }
+    // 2. Query the current user
+    const currentUser = await ctx.db.query.user (
+      {
+        where: {
+          id: ctx.request.userId,
+        }
+      },
+      info
+    );
+    // 3. Check if they have permissions to do this
+    hasPermission(currentUser, ['ADMIN', 'PERMISSIONUPDATE']);
+    // 4. Update the permissions
+    return ctx.db.mutation.updateUser({
+      data: {
+        permissions: {
+          set: args.permissions,
+        }
+      },
+      where: {
+        id: args.userId
+      }
+    }, info);
+  }
 };
 
 module.exports = Mutation;
